requirement.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
Django==4.0.1
django-bootstrap4==21.2
django-filter==21.1
djangorestframework==3.13.1
djongo==1.3.6
haversine==2.5.1
mongoengine==0.23.1
Pillow==9.0.0
pymongo==3.12.3
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.2.4
tzdata==2021.5
